+++ src/index.html
@@ -26,5 +26,7 @@
     </section>
 
     <script type="module" src="index.js"></script>
+    <script type="module" src="."></script>
+    <script type="module" src="./modules/tudo.js"></script>
 </body>
 </html>
