+++ webpack.config.js
@@ -33,4 +33,8 @@ module.exports = {
       },
     ],
   },
+  resolve: {
+    extensions: ['.js'],
+    modules: [path.resolve(__dirname, 'src'), 'node_modules'],
+  },
 };
